requirements.txt
requests
python-dotenv
chromadb
sentence-transformers
pypdf
python-docx
beautifulsoup4
openpyxl